+++ src/Core/Client/Adapter/Http.php
@@ -5,14 +5,15 @@ namespace Solarium\Core\Client\Adapter;
 use Solarium\Core\Client\Endpoint;
 use Solarium\Core\Client\Request;
 use Solarium\Core\Client\Response;
-use Solarium\Core\Configurable;
 use Solarium\Exception\HttpException;
 
 /**
  * Basic HTTP adapter using a stream.
  */
-class Http extends Configurable implements AdapterInterface
+class Http implements AdapterInterface, TimeoutAwareInterface
 {
+    use TimeoutAwareTrait;
+
     /**
      * Handle Solr communication.
      *
@@ -68,7 +69,7 @@ class Http extends Configurable implements AdapterInterface
         $context = stream_context_create(
             ['http' => [
                     'method' => $method,
-                    'timeout' => $endpoint->getTimeout(),
+                    'timeout' => $this->timeout,
                     'protocol_version' => 1.0,
                     'user_agent' => 'Solarium Http Adapter',
                 ],