+++ tests/QueryType/Select/Query/Component/Facet/FacetTest.php
@@ -20,15 +20,15 @@ class FacetTest extends TestCase
 
     public function testConfigMode()
     {
-        $this->facet->setOptions(['key' => 'myKey', 'exclude' => ['e1', 'e2']]);
+        $this->facet->setOptions(['local_key' => 'myKey', 'local_exclude' => ['e1', 'e2']]);
         $this->assertSame('myKey', $this->facet->getKey());
-        $this->assertEquals(['e1', 'e2'], $this->facet->getExcludes());
+        $this->assertEquals(['e1', 'e2'], $this->facet->getLocalParameters()->getExcludes());
     }
 
     public function testConfigModeWithSingleValueExclude()
     {
-        $this->facet->setOptions(['exclude' => 'e1']);
-        $this->assertEquals(['e1'], $this->facet->getExcludes());
+        $this->facet->setOptions(['local_exclude' => 'e1']);
+        $this->assertEquals(['e1'], $this->facet->getLocalParameters()->getExcludes());
     }
 
     public function testSetAndGetKey()
@@ -39,35 +39,28 @@ class FacetTest extends TestCase
 
     public function testAddExclude()
     {
-        $this->facet->addExclude('e1');
-        $this->assertEquals(['e1'], $this->facet->getExcludes());
+        $this->facet->getLocalParameters()->setExclude('e1');
+        $this->assertEquals(['e1'], $this->facet->getLocalParameters()->getExcludes());
     }
 
     public function testAddExcludes()
     {
-        $this->facet->addExcludes(['e1', 'e2']);
-        $this->assertEquals(['e1', 'e2'], $this->facet->getExcludes());
+        $this->facet->getLocalParameters()->addExcludes(['e1', 'e2']);
+        $this->assertEquals(['e1', 'e2'], $this->facet->getLocalParameters()->getExcludes());
     }
 
     public function testRemoveExclude()
     {
-        $this->facet->addExcludes(['e1', 'e2']);
-        $this->facet->removeExclude('e1');
-        $this->assertEquals(['e2'], $this->facet->getExcludes());
+        $this->facet->getLocalParameters()->addExcludes(['e1', 'e2']);
+        $this->facet->getLocalParameters()->removeExclude('e1');
+        $this->assertEquals(['e2'], $this->facet->getLocalParameters()->getExcludes());
     }
 
     public function testClearExcludes()
     {
-        $this->facet->addExcludes(['e1', 'e2']);
-        $this->facet->clearExcludes();
-        $this->assertEquals([], $this->facet->getExcludes());
-    }
-
-    public function testSetExcludes()
-    {
-        $this->facet->addExcludes(['e1', 'e2']);
-        $this->facet->setExcludes(['e3', 'e4']);
-        $this->assertEquals(['e3', 'e4'], $this->facet->getExcludes());
+        $this->facet->getLocalParameters()->addExcludes(['e1', 'e2']);
+        $this->facet->getLocalParameters()->clearExcludes();
+        $this->assertEquals([], $this->facet->getLocalParameters()->getExcludes());
     }
 }