+++ tests/QueryType/Select/Result/Grouping/QueryGroupTest.php
@@ -74,6 +74,6 @@ class QueryGroupTest extends TestCase
 
     public function testCount()
     {
-        $this->assertSame(count($this->items), count($this->group));
+        $this->assertCount(count($this->items), $this->group);
     }
 }